feat(Participants): add Players to teams
Also migrated some dialog component to their own files
